Add new abstract classes for resolving decorators and on() methods
