+++ Source/projections/Builder/OnMethodSpecification.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+import { ProjectionCallback } from '..';
+import { KeySelectorBuilderCallback } from './KeySelectorBuilderCallback';
+import { TypeOrEventType } from './TypeOrEventType';
+
+export type OnMethodSpecification = [TypeOrEventType, KeySelectorBuilderCallback, ProjectionCallback<any>];
