+++ Source/projections/Builder/ProjectionBuilderForReadModel.ts
@@ -17,16 +17,14 @@ import { ProjectionProcessor } from '../Internal';
 
 import { ICanBuildAndRegisterAProjection } from './ICanBuildAndRegisterAProjection';
 import { KeySelectorBuilder } from './KeySelectorBuilder';
-import { KeySelectorBuilderCallback } from './KeySelectorBuilderCallback';
+import { OnMethodBuilder } from './OnMethodBuilder';
+import { TypeOrEventType } from './TypeOrEventType';
 
-type TypeOrEventType = Constructor<any> | EventType;
-type OnMethodSpecification = [TypeOrEventType, KeySelectorBuilderCallback, ProjectionCallback<any>];
 
 /**
  * Represents a builder for building {@link IProjection}.
  */
-export class ProjectionBuilderForReadModel<T> implements ICanBuildAndRegisterAProjection {
-    private _onMethods: OnMethodSpecification[] = [];
+export class ProjectionBuilderForReadModel<T> extends OnMethodBuilder<T> implements ICanBuildAndRegisterAProjection {
 
     /**
      * Initializes a new instance of {@link ProjectionBuilder}.
@@ -35,7 +33,9 @@ export class ProjectionBuilderForReadModel<T> implements ICanBuildAndRegisterAPr
     constructor(
         private _projectionId: ProjectionId,
         private _readModelTypeOrInstance: Constructor<T> | T,
-        private _scopeId: ScopeId) { }
+        private _scopeId: ScopeId) {
+            super();
+        }
 
     /**
      * Defines the projection to operate in a specific {@link ScopeId}.
@@ -45,73 +45,6 @@ export class ProjectionBuilderForReadModel<T> implements ICanBuildAndRegisterAPr
     inScope(scopeId: ScopeId | Guid | string): ProjectionBuilderForReadModel<T> {
         this._scopeId = ScopeId.from(scopeId);
         return this;
-    }
-
-    /**
-     * Add an on method for handling the event.
-     * @template U Type of event.
-     * @param {Constructor<U>} type The type of event.
-     * @param {KeySelectorBuilderCallback<U>} keySelectorCallback Callback for building key selector.
-     * @param {ProjectionCallback<T,U>} callback Callback to call for each event.
-     * @returns {ProjectionBuilderForReadModel<T>}
-     */
-    on<U>(type: Constructor<U>, keySelectorCallback: KeySelectorBuilderCallback<U>, callback: ProjectionCallback<T, U>): ProjectionBuilderForReadModel<T>;
-    /**
-     * Add an on method for handling the event.
-     * @param {EventType} eventType The identifier of the event.
-     * @param {KeySelectorBuilderCallback<U>} keySelectorCallback Callback for building key selector.
-     * @param {ProjectionCallback<T>} callback Callback to call for each event.
-     * @returns {ProjectionBuilderForReadModel<T>}
-     */
-    on(eventType: EventType, keySelectorCallback: KeySelectorBuilderCallback, callback: ProjectionCallback<T>): ProjectionBuilderForReadModel<T>;
-    /**
-     * Add an on method for handling the event.
-     * @param {EventTypeId|Guid|string} eventType The identifier of the event.
-     * @param {KeySelectorBuilderCallback<U>} keySelectorCallback Callback for building key selector.
-     * @param {ProjectionCallback<T>} callback Callback to call for each event.
-     * @returns {ProjectionBuilderForReadModel<T>}
-     */
-    on(eventTypeId: EventTypeId | Guid | string, keySelectorCallback: KeySelectorBuilderCallback, callback: ProjectionCallback<T>): ProjectionBuilderForReadModel<T>;
-    /**
-     * Add an on method for handling the event.
-     * @param {EventTypeId | Guid | string} eventType The identifier of the event.
-     * @param {Generation | number} generation The generation of the event type.
-     * @param {KeySelectorBuilderCallback<U>} keySelectorCallback Callback for building key selector.
-     * @param {ProjectionCallback<T>} method Callback to call for each event.
-     * @returns {ProjectionBuilderForReadModel<T>}
-     */
-    on(eventTypeId: EventTypeId | Guid | string, generation: Generation | number, keySelectorCallback: KeySelectorBuilderCallback, callback: ProjectionCallback<T>): ProjectionBuilderForReadModel<T>;
-    on<U>(typeOrEventTypeOrId: Constructor<T> | EventType | EventTypeId | Guid | string,
-        keySelectorCallbackOrGeneration: KeySelectorBuilderCallback<U> | Generation | number,
-        keySelectorCallbackOrCallback?: KeySelectorBuilderCallback<U> | ProjectionCallback<T>,
-        maybeCallback?: ProjectionCallback<T, U>): ProjectionBuilderForReadModel<T> {
-
-        const typeOrEventType = this.getTypeEventTypeFrom(typeOrEventTypeOrId, keySelectorCallbackOrGeneration);
-        const keySelectorCallback = typeof keySelectorCallbackOrGeneration === 'function'
-            ? keySelectorCallbackOrGeneration
-            : keySelectorCallbackOrCallback as KeySelectorBuilderCallback<U>;
-        const callback = maybeCallback || keySelectorCallbackOrCallback as ProjectionCallback<T, U>;
-
-        this._onMethods.push([typeOrEventType, keySelectorCallback, callback]);
-
-        return this;
-    }
-
-    private getTypeEventTypeFrom<U>(typeOrEventTypeOrId: Constructor<T> | EventType | EventTypeId | Guid | string,
-        keySelectorCallbackOrGeneration: KeySelectorBuilderCallback<U> | Generation | number): Constructor<T> | EventType {
-
-        if (typeof typeOrEventTypeOrId === 'function') {
-            return typeOrEventTypeOrId;
-        }
-
-        if (typeOrEventTypeOrId instanceof EventType) {
-            return typeOrEventTypeOrId;
-        }
-
-        const eventTypeId = typeOrEventTypeOrId;
-        const eventTypeGeneration = typeof keySelectorCallbackOrGeneration === 'function' ? Generation.first : keySelectorCallbackOrGeneration;
-
-        return new EventType(EventTypeId.from(eventTypeId), Generation.from(eventTypeGeneration));
     }
 
     /** @inheritdoc */
@@ -125,13 +58,13 @@ export class ProjectionBuilderForReadModel<T> implements ICanBuildAndRegisterAPr
         cancellation: Cancellation): void {
 
         const events = new EventTypeMap<[ProjectionCallback<T>, KeySelector]>();
-        if (this._onMethods.length < 1) {
+        if (this.onMethods.length < 1) {
             logger.warn(`Failed to register projection ${this._projectionId}. No on methods are configured`);
             return;
         }
-        const allMethodsBuilt = this.tryAddProjectionMethods(eventTypes, events, logger);
+        const allMethodsBuilt = this.tryAddOnMethods(eventTypes, events);
         if (!allMethodsBuilt) {
-            logger.warn(`Failed to register projection ${this._projectionId}. Could not build projection.`);
+            logger.warn(`Failed to register projection ${this._projectionId}. Could not build projection. Maybe it tries to handle the same type of event twice?`);
             return;
         }
         const projection = new Projection<T>(this._projectionId, this._readModelTypeOrInstance, this._scopeId, events);
@@ -145,30 +78,4 @@ export class ProjectionBuilderForReadModel<T> implements ICanBuildAndRegisterAPr
             ), cancellation);
     }
 
-    private tryAddProjectionMethods(
-        eventTypes: IEventTypes,
-        events: EventTypeMap<[ProjectionCallback<any>, KeySelector]>,
-        logger: Logger): boolean {
-        let allMethodsValid = true;
-        const keySelectorBuilder = new KeySelectorBuilder();
-        for (const [typeOrEventTypeOrId, keySelectorBuilderCallback, method] of this._onMethods) {
-            const eventType = this.getEventType(typeOrEventTypeOrId, eventTypes);
-            if (events.has(eventType)) {
-                allMethodsValid = false;
-                logger.warn(`Projection ${this._projectionId} already handles event with event type ${eventType}`);
-            }
-            events.set(eventType, [method, keySelectorBuilderCallback(keySelectorBuilder)]);
-        }
-        return allMethodsValid;
-    }
-
-    private getEventType(typeOrEventTypeOrId: TypeOrEventType, eventTypes: IEventTypes): EventType {
-        let eventType: EventType;
-        if (typeOrEventTypeOrId instanceof EventType) {
-            eventType = typeOrEventTypeOrId;
-        } else {
-            eventType = eventTypes.getFor(typeOrEventTypeOrId);
-        }
-        return eventType;
-    }
 }
