+++ Source/projections/Builder/ProjectionClassBuilder.ts
@@ -12,21 +12,22 @@ import { Constructor } from '@dolittle/types';
 
 import { ProjectionsClient } from '@dolittle/runtime.contracts/Events.Processing/Projections_grpc_pb';
 
-import { IProjections, ProjectionCallback, Projection, KeySelector, DeleteReadModelInstance } from '..';
+import { IProjections, ProjectionCallback, Projection, KeySelector } from '..';
 import { ProjectionProcessor } from '../Internal';
 
 import { CannotRegisterProjectionThatIsNotAClass } from './CannotRegisterProjectionThatIsNotAClass';
 import { ICanBuildAndRegisterAProjection } from './ICanBuildAndRegisterAProjection';
 import { on as onDecorator } from './onDecorator';
-import { OnDecoratedMethod } from './OnDecoratedMethod';
 import { OnDecoratedMethods } from './OnDecoratedMethods';
 import { projection as projectionDecorator } from './projectionDecorator';
 import { ProjectionDecoratedTypes } from './ProjectionDecoratedTypes';
+import { OnDecoratorResolver } from './OnDecoratorResolver';
 
-export class ProjectionClassBuilder<T> implements ICanBuildAndRegisterAProjection {
+export class ProjectionClassBuilder<T> extends OnDecoratorResolver implements ICanBuildAndRegisterAProjection {
     private readonly _projectionType: Constructor<T>;
 
     constructor(typeOrInstance: Constructor<T> | T) {
+        super();
         if (typeOrInstance instanceof Function) {
             this._projectionType = typeOrInstance;
 
@@ -54,17 +55,11 @@ export class ProjectionClassBuilder<T> implements ICanBuildAndRegisterAProjectio
             logger.warn(`The projection class ${this._projectionType.name} must be decorated with an @${projectionDecorator.name} decorator`);
             return;
         }
-
         logger.debug(`Building projection ${decoratedType.projectionId} processing events in scope ${decoratedType.scopeId} from type ${this._projectionType.name}`);
 
-        const methods = OnDecoratedMethods.methodsPerProjection.get(this._projectionType);
-        if (methods === undefined) {
-            logger.warn(`There are no on methods specified in projection ${this._projectionType.name}. An projection must to be decorated with @${onDecorator.name}`);
-            return;
-        }
         const events = new EventTypeMap<[ProjectionCallback<T>, KeySelector]>();
-        if (!this.tryAddAllProjectionMethods(events, methods, eventTypes, logger)) {
-            logger.warn(`Could not create projection ${this._projectionType.name} because it contains invalid projection methods`);
+        if (!this.tryAddAllOnMethods(events, this._projectionType, eventTypes)) {
+            logger.warn(`Could not create projection ${this._projectionType.name} because it contains invalid projection methods. Maybe you have multiple @on methods handling the same event type?`);
             return;
         }
         const projection = new Projection<T>(decoratedType.projectionId, decoratedType.type, decoratedType.scopeId, events);
@@ -76,60 +71,5 @@ export class ProjectionClassBuilder<T> implements ICanBuildAndRegisterAProjectio
                 eventTypes,
                 logger
             ), cancellation);
-    }
-
-
-
-    private tryAddAllProjectionMethods(events: EventTypeMap<[ProjectionCallback<any>, KeySelector]>, methods: OnDecoratedMethod[], eventTypes: IEventTypes, logger: Logger): boolean {
-        let allMethodsValid = true;
-        for (const method of methods) {
-            const [hasEventType, eventType] = this.tryGetEventTypeFromMethod(method, eventTypes);
-
-            if (!hasEventType) {
-                allMethodsValid = false;
-                logger.warn(`Could not create projection method ${method.name} in projection ${this._projectionType.name} because it is not associated to an event type`);
-                continue;
-            }
-
-            const onMethod = this.createProjectionMethod(method);
-            const keySelector = method.keySelector;
-
-            if (events.has(eventType!)) {
-                allMethodsValid = false;
-                logger.warn(`Event handler ${this._projectionType.name} has multiple projection methods handling event type ${eventType}`);
-                continue;
-            }
-            events.set(eventType!, [onMethod, keySelector]);
-        }
-        return allMethodsValid;
-    }
-
-    private createProjectionMethod(method: OnDecoratedMethod): ProjectionCallback<any> {
-        return (instance, event, projectionContext) => {
-            const result = method.method.call(instance, event, projectionContext);
-            if (result instanceof DeleteReadModelInstance) {
-                return result;
-            }
-            return instance;
-        };
-    }
-
-    private tryGetEventTypeFromMethod(method: OnDecoratedMethod, eventTypes: IEventTypes): [boolean, EventType | undefined] {
-        if (this.eventTypeIsId(method.eventTypeOrId)) {
-            return [
-                true,
-                new EventType(
-                    EventTypeId.from(method.eventTypeOrId),
-                    method.generation ? Generation.from(method.generation) : Generation.first)
-            ];
-        } else if (!eventTypes.hasFor(method.eventTypeOrId)) {
-            return [false, undefined];
-        } else {
-            return [true, eventTypes.getFor(method.eventTypeOrId)];
-        }
-    }
-
-    private eventTypeIsId(eventTypeOrId: Constructor<any> | EventTypeId | Guid | string): eventTypeOrId is EventTypeId | Guid | string {
-        return eventTypeOrId instanceof EventTypeId || eventTypeOrId instanceof Guid || typeof eventTypeOrId === 'string';
     }
 }
