+++ Source/projections/Builder/TypeOrEventType.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+import { Constructor } from '@dolittle/types';
+import { EventType } from '@dolittle/sdk.events';
+
+export type TypeOrEventType = Constructor<any> | EventType;
