+++ Source/projections/Builder/index.ts
@@ -9,6 +9,9 @@ export { KeySelectorBuilderCallback } from './KeySelectorBuilderCallback';
 export { OnDecoratedMethod } from './OnDecoratedMethod';
 export { OnDecoratedMethods } from './OnDecoratedMethods';
 export { on } from './onDecorator';
+export { OnMethodSpecification } from './OnMethodSpecification';
+export { OnMethodBuilder } from './OnMethodBuilder';
+export { OnDecoratorResolver } from './OnDecoratorResolver';
 export { ProjectionBuilder } from './ProjectionBuilder';
 export { ProjectionBuilderForReadModel } from './ProjectionBuilderForReadModel';
 export { ProjectionClassBuilder } from './ProjectionClassBuilder';
@@ -20,3 +23,4 @@ export { ProjectionOptions } from './ProjectionOptions';
 export { ProjectionsBuilder } from './ProjectionsBuilder';
 export { ProjectionsBuilderCallback } from './ProjectionsBuilderCallback';
 export { ReadModelAlreadyDefinedForProjection } from './ReadModelAlreadyDefinedForProjection';
+export { TypeOrEventType } from './TypeOrEventType';
